Prepare for new closure determination
The plan is to not use the parser to determine the closure
block of nested functions, but to instead use the callgraph
visitor class. This should cope better with some of the currently
failing scenarios in fay.pas for example, where we don't discover
the usage for recursive calls to outer nested functions because the
code scans the usage before the outer function is complete.

Since the callgraph is performed on the final AST, all functions
are already complete.

This commit paves the way without breaking current code, and proviides
some of the groundwork needed to progress.

Clean up some codegen functions.

Clean up PrototypeAST, and remember the function created.
Change name from ProtoTypeAST::CodeGen to ProtoTypeAST::Create.

Replace some useles sequences of `if (cond) assert(0 && ...);` with just
`assert(!cond && ...)`.