Update makefile
Disable warnings for unused parameters.

Change the way llvmversion file is generated to use only llvm resources.
This does assume that the version used is actually the HEAD of the source
of the LLVM source directory (and that the source is available on the machine).
This avoids using clang to get the version, which may not be available, or the
same version that the LLVM version is.