Fix By Value argument passing
The code needs to add the type to the attribute for by value arguments, rather than
just using the ByVal attribute itself.

This fixes 11 test failures on Arm64, as the compiler crashes during the llvm code
generation. It is not noticeable on x86, because the type is only retrieved but
not actually used - so it doesn't matter that it is NULL.