Refactor argument list processing.
Split out argument processing to make functions a bit smaller and "do one thing".

This is preparation for modifying the trampoline code.

Also move the code for detecting "non-variable passed to a var parameter" and make
a better error message for that case.

Introduce IsClosure() function and related constructor argument to VarDef.