Fix some set operations
On assignments, set the right hand type of the assignment before running
the binary expression checking on the right-hand side.

This meant the typecast on the right hand side needs to be visited too, or the
compiler crashes on 2 * x where x is a real (for having different types on lhs
and rhs).