Refactor integer type declaratons.
This is the beginning of supporting more sizes of integer types.
Use template design to produce integers in 32- and 64-bit lengths,
unifying most of the functionality.

Adding read function for int64 types, as that was missing.

Fix some issues with range declarations - again in preparation
for supporting shorter storage forms.