Add more complex functionality: support for Arg and fix bug in Abs
This adds the Arg function (which is basically just an atan2 - we may want
to simply call atan2 instead, but for now, this'll do.

The Abs and Arg functons return a real from complex input. The test I was working
with was incorrectly caling print_complex from the Abs function.