Fix complex math in 32-bit mode
Calling convention changes between 64-bit and 32-bit mode, so make it the
same always to avoid having checks in the compiler.

Also fix latent bug that the <> operator in test was actually using the =
operator in the test - and of course, then checking the wrong values in
the expected result.