[NFC]Don't look in list for symbols we found in lexer
When we have identified the symbol in the switch statement, assing
the token type directly, instead of searching through a list. Probably
not a big gain, but there's no reason to go on and loop over the
symbols if we know the answer.

Also unify style covering multiple choices with switch, not a mix
of if and switch. This also avoids an extra call to PeekChar if the
first choice was not a match.