Change from sscanf to strto{l,ll,d} for val
The implementation of val was using sscanf to convert string to
numeric form. This is quite inefficient. Intead, use the strtol,
strtoll and strtod functions to do the same work.

When comparing the time, this is about 2.4 times faster (although
unlikely to matter that much for anything other than my silly
benchmark that does 10 million conversions of 123.45).

Fix strings to have a spare space for a zero at the end, and set
it to zero - this was a latent bug by all account.