Fix bug where same name exists in different scope
When generating closure for a function, the code used the name
of the function to determine what function to apply what closure
to. This doesn't work if there are multiple functions with the same
name in different scopes, as they will have different closures.

Fix is to compare the pointer to the actual function instead.

Also found that the code was calling AddClosureArg in the parser, when
the closure is never present there, so a waste of time.