Error out when assigning to constants
Assigning to constants records and arrays would happily succeed to
compile, but not generate correct code (it wouldn't actually update
the value, leading to incorrect result.

This patch introduces a new function to search for a base
expression in the AST, and use that to determine if
the base of an assignment is a variable. If it isn't, then
fail with an error message. This fixes the above condition
when trying to assign to a constant.

Also had to add some missing accept functions and classof
member functions.

New test added to check that the compiler gives an error.