[NFC] Do not make so many copies of tokens.
Pass token by const reference whenever possible (most places).

Simplify the error output so that we don't pass CurrentToken, but
just pass the error message, and pick up CurrentToken() in the
print function.

This saves a bunch of copy operations, and reduces the code size by
nearly 6KB on x86-64.

More importantly, it makes the code easier to read in most places.