Record name of builtin function to make nicer semantic errors.
The error messages for  semantic check of bultin functions is pretty poor.
This change allows us to print the name of the function as a starting point.

There's plenty of room for further improvement, but this is a step in the
right direction.

Some trivial refactoring also made.

No functional change intended for the compiled code generated.