Fix broken code
Running ISO10206 tests showed a couple of bugs had crept in with the past
few changes.

This fixes those and adds tests to avoid those things being broken again.
Note that the bind test currently doesn't actually work. It is still
not clear to me exactly how bind is meant to work.

Also added a line to the improvement-todo.txt to remind myself that the
error recovery definitely should be improved, so there isn't a lot of error
messages from one small mistake.