Semantics: Check that condition on while, if and repeat is boolean
It is an error to use a non-boolean expression as the condition
in while, repeat .. until and if. Previously this would lead
to a crash in LLVM. Now we show a diagnostic message, which
makes it easier to understand what is wrong.

Also fixed a few missing "visit" calls for the visitor pattern,
including the ones for repeat and while.

Generated code for valid input source code is unchanged.