Move constant function evaluation to constants.cpp
The evaluation and creation of new constants from constant functions
is more to do with constant handling than parsing, so move it into
the source file that deals with constants.

This also moves a couple of other functions, and making a helper function
that identifies constant functions, and moving the ConstDeclToInt function
into the same file.

No functional change intended.