Reduce number of boiler-plate virtual functions on types
Make separate, rather the virtual member functions for some of the
trivial boolean values and such.

Some functions were not really used in a particularly useful way, so
could be completely removed and replaced - Bits for example can be
substituted with a call to Size() * CHAR_BIT.

This reduces the code size by some fair amount, and I think it makes
it easier to follow too.

No functional change intended.