Add very simple schema parsing.
This supports the "template" system in ISO-10206 Pascal, but only
for the most simple of use-case: range with a constant start and
a variable end value.

This is by no means a functional implementation, but committed
as a basic starting point, to avoid having a very big change
all at once.