Support character array for trim operations.
This is of course what it was originally designed to be used for, but the
previous implementation was a bit broken.

New test added in the newstringfunctions test file.

A bug was also found in comparing character arrays. This was fixed
at the same time, and a new regression test added