Support different size index values in different parts of an array
Multidimensional arrays need to use the same index type in the flattened
1D array is ultimately generated.

This finds the largest type and uses that for the index type, rather than
the type of the type of the individual ranges describing the index. This
means LLVM won't assert when trying to multiply a constant of type i8 by a
constant of type i32.

No real functional change, just unifying types.