[NFC] Remove uses of true/false as arguments.
Use enums instead of booleans. In one case, pass in a type declaration
instead of boolean.

This is just to avoid having smatterings of true/false that determine
the path through a function, and make it a bit clearer what that extra
argument does.

This is not changing any functionality in the compiler, just tidying
things up.